24 points Eliezer_Yudkowsky 08 September 2012 08:41:56AM
The following quotes were heavily upvoted, but then turned out to be made
by a Will Newsome sockpuppet who edited the quote afterward. The original
comments have been banned. The quotes are as follows:

    If dying after a billion years doesn't sound sad to you, it's because
    you lack a thousand-year-old brain that can make trillion-year plans.

-- Aristosophy

    One wish can achieve as much as you want. What the genie is really
    offering is three rounds of feedback.

-- Aristosophy

If anyone objects to this policy response, please PM me so as to not feed
the troll.